+++ blockify/blocklist.py
@@ -17,26 +17,26 @@ class Blocklist(list):
         self.location = util.BLOCKLIST_FILE
         self.use_substring_search = util.CONFIG["general"]["substring_search"]
         self.extend(self.load())
-        log.info("Blocklist loaded from {}.".format(self.location))
+        log.info(f"Blocklist loaded from {self.location}.")
         self.timestamp = self.get_timestamp()
 
     def append(self, item):
         "Overloading list.append to automatically save the list to a file."
         # Only allow nonempty strings.
         if item in self or not item or item == " ":
-            log.debug("Not adding empty or duplicate item: {}.".format(item))
+            log.debug(f"Not adding empty or duplicate item: {item}.")
             return
-        log.info("Adding {} to {}.".format(item, self.location))
+        log.debug(f"Adding {item} to {self.location}.")
         super(Blocklist, self).append(item)
         self.save()
 
     def remove(self, item):
-        log.info("Removing {} from {}.".format(item, self.location))
+        log.debug(f"Removing {item} from {self.location}.")
         try:
             super(Blocklist, self).remove(item)
             self.save()
         except ValueError as e:
-            log.warning("Could not remove {} from blocklist: {}".format(item, e))
+            log.error(f"Could not remove {item} from blocklist: {e}")
 
     def find(self, song):
         if self.use_substring_search:
@@ -66,7 +66,7 @@ class Blocklist(list):
         return [i for i in blocklist.split("\n") if i]
 
     def save(self):
-        log.debug("Saving blocklist to {}.".format(self.location))
+        log.debug(f"Saving blocklist to {self.location}.")
         with codecs.open(self.location, "w", encoding="utf-8") as f:
             f.write("\n".join(self) + "\n")
         self.timestamp = self.get_timestamp()